% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derandomKnock.R
\name{derandomKnock}
\alias{derandomKnock}
\title{A stable variable procedure based on the knockoffs}
\usage{
derandomKnock(
  X,
  y,
  type = "pfer",
  M = 30,
  tau = 0.5,
  seed = 24601,
  v = 1,
  k = 1,
  alpha = 0.05,
  knockoff_method = "gaussian",
  mu = NULL,
  Sigma = NULL,
  pInit = NULL,
  Q = NULL,
  pEmit = NULL
)
}
\arguments{
\item{X}{a n-by-p matrix of the covariates.}

\item{y}{the response vector of length in (can be continuous or binary).}

\item{type}{the type of error to control. Options include "pfer" and "kfwer".}

\item{M}{an integer specifying the number of knockoff copies computed (default: 30).}

\item{tau}{a number betweem 0 and 1 indicating the selection frequency (default: 0.5).}

\item{seed}{an integer specifying the random seed used in the procedure.}

\item{v}{a positive numver indicating the PFER target (default: 1). Can be left NULL if using the kfwer error.}

\item{k}{a positive integer corresponding to k-FWER.}

\item{alpha}{a number between 0 and 1 indicating the target k-FWER level.}

\item{knockoff_method}{either "gaussian" or "hmm".}
}
\description{
The main function the implement the derandomized knockoffs procedure.
}
