% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_lp.R
\name{get_lp_bnd}
\alias{get_lp_bnd}
\title{Determine the parameter for PFER/k-FWER control}
\usage{
get_lp_bnd(M, eta, beta = 1)
}
\arguments{
\item{M}{an integer specifying the number of knockoff copies computed (default: 30).}

\item{eta}{a number betweem 0 and 1 indicating the selection frequency (default: 0.5).}

\item{beta}{a number between 0 and 1 specifying the vanishing speed of the density.}
}
\description{
Determine the parameter for PFER/k-FWER control
}
