% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameter.R
\name{get_params}
\alias{get_params}
\title{Determining the parameter for the base filter}
\usage{
get_params(
  type,
  v = NULL,
  k = NULL,
  alpha = NULL,
  M = NULL,
  eta = NULL,
  thres = 50,
  beta = 1
)
}
\arguments{
\item{type}{the type of error to control. Options include "pfer" and "kfwer".}

\item{v}{a positive numver indicating the PFER target (default: 1). Can be left NULL if using the kfwer error.}

\item{k}{a positive integer corresponding to k-FWER.}

\item{alpha}{a number between 0 and 1 indicating the target k-FWER level.}

\item{M}{an integer specifying the number of knockoff copies computed (default: 30).}

\item{eta}{a number betweem 0 and 1 indicating the selection frequency (default: 0.5).}
}
\description{
A function to determine the parameters for the derandomized knockoffs procedure
}
