% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getV.R
\name{getV}
\alias{getV}
\title{Stable knockoff(fdx)
internal function to determine v}
\usage{
getV(k, alpha, accuracy = 0.05, tau = 0.5, h1, h2, nu, xi)
}
\description{
Stable knockoff(fdx)
internal function to determine v
}
