% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfer_filter.R
\name{pfer_filter}
\alias{pfer_filter}
\title{A filter}
\usage{
pfer_filter(
  X,
  y,
  v0 = 1,
  M = 50,
  tau = 0.5,
  knockoff_method = "gaussian",
  knockoff_stat = stat.glmnet_coefdiff,
  seed = 24601,
  mu = NULL,
  Sigma = NULL,
  pInit = NULL,
  Q = NULL,
  pEmit = NULL
)
}
\description{
A filter
}
