% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getV.R
\name{getV}
\alias{getV}
\title{Cumputing the parameter of the base procedure}
\usage{
getV(k, alpha, accuracy = 0.05, tau = 0.5, h1, h2, nu, xi)
}
\description{
internal function to determine the parameter of the base procedure
}
